[PROD-910] Loosen validation of Google appengine operation ID format (#1557)
While historically Google's operation ID has consistently been returned to us as a UUID, that was not guaranteed.
We are now seeing IDs returned in different formats (e.g. operation-1702658910061-60c8f2b820248-67c4276a-e6568941 which has 'operation-' prefix and a suffix that does not conform to the UUID format).
This was causing issues for tests and users when creating certain resources (though not consistently).