[DR-3423] DrsService parallel streams converted to use Executor
Parallel streams use the default ForkJoinPool.commonPool, which by default has one less threads as you have processors.
They are also not recommended for IO-bound tasks.
This PR refactors out our three parallel streams in favor of using a new ThreadPoolTaskExecutor dedicated to DRS resolution operations.
It is set up as a bean in new DrsConfiguration record, in which I've pulled in an existing configuration for max DRS lookups.
Using executors to manage parallel execution is an established pattern in TDR, but here I've created a ThreadPoolTaskExecutor rather than a ThreadPoolExecutor as we do elsewhere.
Micrometer will automatically monitor ThreadPoolTaskExecutors for us without additional code, and we'll be able to visualize the state of such a threadpool in Grafana.
Added unit tests and removed an integration test whose functionality is now tested by unit tests.