[DR-3409] Handle large string, text and array fields in Azure ingest (#1568)
* Drive by fix to make tests pass locally

This was causing issues with bean resolution when default google credentials are not a service account (e.g. when you are logged in with a Google user account).

* Make our use of Synapse handle text that is larger than 8000 bytes

In two cases:
- When there are straight up more than 8K bytes in a text field
- When the serialized JSON is more than 8K bytes

Note: this makes it so that the default column handling on ingest is to explicitly
cast the JSON values. The requiresJSONCast option on columns is repurposed to indicate
loading the json value using OPENJSON

* Unfortunately, varchar(max) doesn't appear to work with CSV files

* Make sure that the special json loading happens for string fields too

* PR Feedback:

- add a test that ingest fails on larger csv fields
- use @Nullable annotation for service account email parameter
- update API doc for csv ingests