FutureUtils isn't thread-safe, obscures root cause exceptions
FutureUtils.waitFor uses an AtomicReference but doesn't use it atomically.
If an ErrorReportException is emitted by a Future of executing, throw it instead of wrapping it in an ApiException (which has a 500 HTTP status and obscures the true problem).
Updated FutureUtilsTest and converted to a pure JUnit 5 unit test (but there are still a few smells in this test class, lots of sleeps).