DC-824: Fix the liquibase plugin config in build.gradle (#1584)
The liquibase gradle plugin isn't used as part of our build, so it's easy for it to stop working without us noticing. This fixes it so it works again. Its main use is as a quick way to test out liquibase changes locally or reset both the datarepo and stairway databases locally.