DC-850: fix issue where datarepo-client subproject build was failing (#1592)
The client subproject build was failing when trying to run the gradle task that pushes the client .jar to artifactory. It turns out that the changes to the project now require JDK 17, and the client subproject was using JDK 11, which is the default JVM in the ubuntu-latest image. Adding a job to explicitly set up JDK 17 fixes this problem.

Note that this wasn't an issue for the main project, as it had already been updated to JDK 17 via a change to the datarepo-actions and doesn't use the java in the ubuntu-latest image. broadinstitute/datarepo-actions#72

Failing build: https://github.com/DataBiosphere/jade-data-repo/actions/runs/7888933476/job/21527561449
Passing build: https://github.com/DataBiosphere/jade-data-repo/actions/runs/7892179015/job/21538097580

[ Merged with Override Branch Protections because changing this file doesn't trigger tests ]