DC-850: Fix container build / bumper github action (#1596)
The previous fix didn't work because the datarepo action to build the docker container failed due to the JDK 17 change. This separates the container build from the client build to prevent this problem. The bumper task was also refactored into its own job, and a new dependency was added so that the sherlock publish now requires the cherry-pick step to have passed.

[ Merged with Override Branch Protections because changing this file doesn't trigger tests ]

#major