[WOR-1551] TDR cannot deserialize Rawls workspace responses
If a snapshot has been exported to a workspace with its policy groups synced, and the workspace is accessible to the TDR user, TDR should be able to resolve the workspace's UUID back to a human-readable representation of the workspace (name and link).
This has worked in the past, but Rawls' workspace fetch response body has expanded since it was first added, and our deserialization was not resilient to those changes.

Converted our workspace response deserialization to use records (the latest prevailing approach in TDR) with unknown properties ignored.
Expanded unit testing throughout.