[WOR-1544] TDR users should contact Terra Support for help
When a TDR Stairway flight fails in an unknown way, we've previously instructed users to contact the TDR team for help.
Now that Terra Support is the first line for user inquiries, we should update this message to avoid unnecessary rerouting of user questions.
Example: https://broadinstitute.slack.com/archives/CD4HBRFMG/p1709145114333539