DC-857: generate hierarchy from database query (#1625)
This change replaces the stub code that implements the `getHierarchy` cohort builder API with one that generates a result using a database query. Note that the values for `count` and `hasChildren` are still stubbed.

Given a concept ID, the query finds all ancestors using the `concept_ancestor` table. Then, for each ancestor it finds all of its children by using the `concept_relationship` table. The result produced is a list of `(parent, children)` pairs where `parent` is the concept ID of the parent node, and `children` is a list of `concept` objects for each child of that parent.

The previous API design used a single root node to return the result, but that approach didn't work because the concept hierarchy is a graph and a concept may have more than one parent.