Add srcclr plugin, and instructions for running scans
This includes running scans off of feature branches via AppSec's GHA as well as running the Gradle srcclr task off of local code.

These tools are useful to verify that you've fully remediated reported vulnerabilities and haven't brought in new ones: the GHA in particular uses the same mechanism and configuration that AppSec uses to flag vulnerabilities to us.