Obtaining Step name from stack trace sidesteps ClassNotFoundExceptions
When a Stairway step throws an exception, we massage it before serializing it and writing it to the database.
This has included walking the stack trace elements to find out what Step threw the exception.
A ClassNotFoundException in this process would previously result in a new exception being thrown, and a dismal flight failure.
But we can legitimately see unknown classes in the stack trace (such as from reflection) which should not prevent our search for the source Step.
Example exception which previously threw on jdk.internal.reflect.GeneratedMethodAccessor980:

Step Result exception
bio.terra.service.load.exception.LoadLockedException: Load 'SM-NKIZ7-24-03-15_16-02-19-SM-NKIZ7.files' is locked by flight 'uEoAG5uQRTaFq4ov_OE-RQ'
	at bio.terra.service.load.LoadDao.lockLoad(LoadDao.java:102)
	at jdk.internal.reflect.GeneratedMethodAccessor980.invoke(Unknown Source)…
	at bio.terra.service.load.LoadService.lockLoad(LoadService.java:35)
	at bio.terra.service.load.flight.LoadLockStep.doStep(LoadLockStep.java:23)