Modify RetrieveDataset test to patch description instead of PHS ID
Our test users will be custodians, and should have permission to patch a dataset's description.  Now that our test users are not automatically added to new datasets' admin policy, they do not have permission to add new stewards.

Previously, this test patched a dataset's PHS ID as a means of testing that PATCH functionality was working following an upgrade to Java 17.  The value of this test is therefore retained, even if there might be another way to achieve this.