[No Ticket] Document Integration test user credentials Vault path (#1650)
We occasionally have needed to log in as an integration test user when integration tests are failing (ex. when a user needed to accept Terms of Service).

Rather than passing around the password in Slack, documenting the Vault path where the Harry Potter user credentials live in our integration test configuration.