[DCJ-28] Use Sherlock to Slack DCJ team on failures, retries (#1654)
* Use Sherlock to Slack DCJ team on failures, retries of dev image update

Previously, this ran on non-broadbot commits to develop branch and emitted a Slack alert to #jade-spam on each run.

Now, Sherlock will only emit a Slack alert to #dsp-data-custodian-journeys on failures or retries (presumably retries run when an initial attempt has failed, so we want to know that they succeed).

This Slack channel is set as Github repository variable SLACK_NOTIFICATION_CHANNELS at https://github.com/DataBiosphere/jade-data-repo/settings/variables/actions

* Use Sherlock to Slack DCJ team on failures, retries of nightly test runs

Previously, this emitted a Slack alert to #jade-alerts on each scheduled run.

Now, Sherlock will only emit a Slack alert to #dsp-data-custodian-journeys on failures or retries (presumably retries run when an initial attempt has failed, so we want to know that they succeed) running of develop branch.

Note that these tests failing are not considered release blockers: they must pass on PRs before merge, so we do not emit notifications to QA for triage.

* Use Sherlock to Slack DCJ team on failures, retries of Python client publish

Previously, this ran on commits to develop branch and emitted a Slack alert to #jade-spam on each run.

Now, Sherlock will only emit a Slack alert to #dsp-data-custodian-journeys on failures or retries (presumably retries run when an initial attempt has failed, so we want to know that they succeed).

Additional notable change: this workflow is best suited to run as a Beehive deployment hook on datarepo-dev, only triggered once after a successful deployment.  I will create this in Beehive after approval and prior to merge.

Presently, it was running on both the merge commit and corresponding broadbot tag bump commit to develop.  I believe this is producing a fencepost error where Python client changes are incorrectly attributed to the preceding semver.

* Use Sherlock to Slack on staging smoke test runs

Previously, this emitted Slack alerts to #dsde-qa and #jade-alerts on each run, triggered via a Beehive deploy hook on datarepo-staging.  We consider these tests a release blocker if they fail.

Now, Sherlock handles the Slack alerts:
- #dsde-qa on every run
- #dsp-data-custodian-journeys on failures or retries (presumably retries run when an initial attempt has failed, so we want to know that they succeed)

* Remove Slacking entirely from dev-image-update.yaml

Instead, I will create a Beehive Slack deploy hook on datarepo-dev that notifies #dsp-data-custodian-journeys and tags the authors of the included commits.

It will be first emitted when a deployment kicks off and be updated over the course of the deployment.

* Whitespace

* Temporarily Slack on PR test runs

To manually demonstrate thee changes.

* Revert "Temporarily Slack on PR test runs"

This reverts commit a6f9a1b98b026aaec6aad5da571aecba9bd97fef.

* Update action name: test run publish doesn't notify