DC-996: fix searchConcepts issue on GCP (#1655)
Fix issue where `searchConcepts` API was broken for GCP datasets.

Previous commit https://github.com/DataBiosphere/jade-data-repo/pull/1652 introduced a bug where `searchConcepts` for GCP projects didn't work, due to the way it represented boolean `true` values. For Azure, `1` must be used, since t-sql doesn't have a boolean type that can be used in a `SELECT`. For GCP `1` doesn't work as it can't be converted to a boolean type.

The code now uses `true` on GCP and `1` on azure, and supports `false` values as well.