Refactor snapshotUnauthorizedPermissionsTest to unit tests
This integration test was failing without the Admins group specified, but it shouldn't have been.
When looking more closely, I saw that it was checking permissions on the wrong users, which was masked by the presence of the Admins group.
The functionality tested was a better fit to refactor to controller unit tests.