[DCJ-15] Make @DataBiosphere/data-custodian-journeys team the default codeowners
And make @DataBiosphere/data-explorer-eng team codeowners of files in snapshotbuilder directories.

I have configured the new DCJ team so that 2 reviews are automatically requested when the team is a codeowner, rotating through reviewers via a load balancer strategy.  Sebastian and Gideon (our contractors) are members of this team but exempt from review requests.