[DCJ-15] Make DCJ team the default codeowners (#1665)
* [DCJ-15] Make @DataBiosphere/data-custodian-journeys team the default codeowners

And make @DataBiosphere/data-explorer-eng team codeowners of files in snapshotbuilder directories.

I have configured the new DCJ team so that 2 reviews are automatically requested when the team is a codeowner, rotating through reviewers via a load balancer strategy.  Sebastian and Gideon (our contractors) are members of this team but exempt from review requests.

* PR feedback: DE team are also codeowners

Their snapshot builder work often touches code outside of /snapshotbuilder directories, and this is likely to be the case for the next 6 - 8 months.

* [docs] @DataBiosphere/broadwrite membership is usually sufficient

A DSP-er wanting to contribute to jade-data-repo is able to do so without needing to be a codeowner or a repository admin.

Our "Getting Started" guide has historically been tailored towards onboarding new members of a team stewarding the repo.