Refactor AzureStorageAuthInfo as a Record
Motivation: new unit tests will need to instantiate this object to pass around to mocks.  The prior implementation did not have an `equals` method defined, so mocking didn't behave as expected.

To eliminate the need to define such a method, eliminate boilerplate, and get human-readable logging of this object, I converted this class to a Record. We've taken on similar conversions throughout the repo for POJOs and config objects.