[DCJ-444] Azure ingest validates filerefs before publishing tabular data (#1712)
* Azure ingest validates filerefs before publishing tabular data

1. Modify IngestValidateAzureRefsStep to validate the filerefs present in the staging table, rather than off of published Parquet files.
2. Rename to IngestValidateScrateTableFilerefsStep for clarity.
3. Reorder flight so that this step comes before IngestCreateParquetFilesStep: this way if an attempted ingest contains invalid file references, TDR will not have published the final Parquet files.  At no point during such a flight or after its rollback would users see the tabular data in their dataset.

* Refactor AzureStorageAuthInfo as a Record

Motivation: new unit tests will need to instantiate this object to pass around to mocks.  The prior implementation did not have an `equals` method defined, so mocking didn't behave as expected.

To eliminate the need to define such a method, eliminate boilerplate, and get human-readable logging of this object, I converted this class to a Record. We've taken on similar conversions throughout the repo for POJOs and config objects.

* Fix Column.equals NPE when comparing ids

Motivation: needed for upcoming unit tests.

A Column may not have an id specified, as is the case with when we convert Columns to SynapseColumns (which extends Column).

Change the id equality check to null-safe Objects.equals method.

* Add unit tests for IngestValidateScratchTableFilerefsStep

Cases tested:

- Ingesting to a table with no fileref columns succeeds (as a no-op)
- Ingesting only valid filerefs succeeds
- Ingesting 1+ invalid filerefs throws an InvalidFileRefException
- Only the invalid filerefs are flagged in the exception's causes

* Add unit test for earlier Column equality bugfix