[DCJ-610] JobModel includes more info on targeted resource
Specifically, iam_resource_type, iam_resource_id, and iam_resource_action.

Motivation: when debugging flights on behalf of users, it's not always clear from a flight's description or log spelunking what resource is being targeted.

We already pass this information as flight inputs for the majority of our flights, to enable Sam-based permission checks when retrieving information on flights.