@pshapiro4broad PR feedback
Notably, a shift in my understanding of "who's locking what":
Flights lock a "load lock key" comprised of a load tag and a dataset ID.
In other words, concurrent load operations (e.g. file ingests) to a dataset are not allowed to use the same load tag.

In an effort to further reduce confusion, the `load` table is now named `load_lock` and has had its `locked` column remove.  The presence of a row in the table denotes that a lock is active.