[DCJ-32] Load tags are locked and unlocked at the dataset level (#1785)
* [DCJ-32] Load tags are locked and unlocked at the dataset level

Let flight1 have an active lock on loadTagA while ingesting files to datasetX.

Previously, if flight2 attempted to lock loadTagA while ingesting files to datasetY (a different dataset), the attempt would fail.  Now, such an attempt would succeed.

Unchanged: if flight2 attempted to lock loadTagA while ingesting files to datasetX (the same dataset), the attempt will fail.

* Address Sonar issues, increase test coverage

* LoadLockStep, LoadUnlockStep get dataset ID from working map

* Fix Sonar issues in IngestUtilsTest

We can use the inputParameters class field instead.

* @pshapiro4broad PR feedback (first pass)

* @pshapiro4broad PR feedback

Notably, a shift in my understanding of "who's locking what":
Flights lock a "load lock key" comprised of a load tag and a dataset ID.
In other words, concurrent load operations (e.g. file ingests) to a dataset are not allowed to use the same load tag.

In an effort to further reduce confusion, the `load` table is now named `load_lock` and has had its `locked` column remove.  The presence of a row in the table denotes that a lock is active.

* LoadLockedException can be thrown for generic load operation conflicts

Not just for file ingests.