LoadUnlockStep accepts user-supplied lockName
The default behavior when unspecified is to unlock the current flight's ID (when clearing a lock taken out earlier in the flight).
In the case of manual dataset unlocking, we want to clear a lock taken out by a past flight.
Manual dataset unlocking will not have the load tag to unlock in its flight context, so unlocks any load tag associated with the dataset ID and lock name.